Add Apollo boilerplate, refactor react app
